屏安柜新增开锁状态和还柜状态2个页面,任务广告接口对接完成

+++ app/src/main/java/com/muju/note/launcher/app/Cabinet/bean/CabinetBean.java
@@ -78,7 +78,7 @@ public class CabinetBean implements Serializable {
         private String cabinetCode;
         private String channel;
         private String createTime;
-        private int deposit;
+        private double deposit;
         private int deptId;
         private int deviceType;
         private int expireFee;
@@ -134,11 +134,11 @@ public class CabinetBean implements Serializable {
             this.createTime = createTime;
         }
 
-        public int getDeposit() {
+        public double getDeposit() {
             return deposit;
         }
 
-        public void setDeposit(int deposit) {
+        public void setDeposit(double deposit) {
             this.deposit = deposit;
         }
